Updates with browserified version of data library
The upstream version of this library was/is intended for use in a
node process. The use of `fs` adds a bit of complexity when building
exclusively for the browser.

For the time being, we will be using a flattened version of this file.

The flattening was accomplished via brfs & browserify.